% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjointSimulator-package.R
\docType{package}
\name{BPTOSimulator-package}
\alias{BPTOSimulator}
\alias{BPTOSimulator-package}
\title{BPTOSimulator: Conjoint Simulator Boilerplate}
\description{
Boilerplate BPTO Simulator based on HB.
}
\author{
\strong{Maintainer}: Michal Czyz \email{michal@codingmanatee.ninja}

}
\keyword{internal}
