% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_calc_sample.R
\name{calc_sample}
\alias{calc_sample}
\title{Calculate Sample Size}
\usage{
calc_sample(respid_key)
}
\arguments{
\item{respid_key}{IDs from data frame with currently applied filters.}
}
\value{
Numeric. Sample size.
}
\description{
Calculates the sample size based on filters.
}
