% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calc_shr.R
\name{calc_scaler}
\alias{calc_scaler}
\title{Map Price To New Range}
\usage{
calc_scaler(price_set, old_max, old_min, new_max = 1, new_min = -1)
}
\arguments{
\item{price_set}{Numeric. Price in current range.}

\item{old_max}{Numeric. Maximum price of product.}

\item{old_min}{Numeric. Minimal price of product.}

\item{new_max}{Numeric. New maximum price of product. Default `1`.}

\item{new_min}{Numeric. New minimum price of prouct. Default `-1`.}
}
\value{
Numeric. Price with the new value that is in the range of `new_max`
  and `new_min`.
}
\description{
Takes a number from range and maps it to new range.
}
