% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_shares_dt_wrapper.R
\name{dt_common_formats}
\alias{dt_common_formats}
\title{DT Formating Wrapper}
\usage{
dt_common_formats(data)
}
\arguments{
\item{data}{DT data frame}
}
\value{
Data frame with filtered records.
}
\description{
A wrapper to format DT table currency and percents.
}
