% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_elasticity_wrappers.R
\name{elasticity_list_vals}
\alias{elasticity_list_vals}
\title{List of Elasticities}
\usage{
elasticity_list_vals(iterated_df)
}
\arguments{
\item{iterated_df}{Data frame which is result of iter shares function.}
}
\value{
List with calculated elasticity.
}
\description{
Calulate elasticity of shares, sales, channel sales and profit.
}
