% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calc_shr.R
\name{iter_share}
\alias{iter_share}
\title{Calulate Shares for Elasticyty Simulation}
\usage{
iter_share(
  product_input,
  product_output,
  data,
  specs,
  respid_key,
  product_names,
  current_prices,
  current_dm,
  current_cost,
  current_ms,
  n_points = 20
)
}
\arguments{
\item{product_input}{Input product for which elasticity will be calculated.}

\item{product_output}{Output product for which elasticity will be calculated.}

\item{data}{Data frame with hb utilities.}

\item{specs}{Data frame with products specification.}

\item{respid_key}{Vector with Respondents ID to filter.}

\item{product_names}{Vector of product names.}

\item{current_prices}{Snapshot of prices from Assumptions tab.}

\item{current_dm}{Snapshot of distribution margin from Assumptions tab.}

\item{current_cost}{Snapshot of cost from Assumptions tab.}

\item{current_ms}{Snapshot of market size from Assumption tab.}

\item{n_points}{Number of points for which Max-Min range will be divided.}
}
\value{
Data frame with calculated shares by price step.
}
\description{
Iterates through prices of Input brand and calulates
shares for Output brand.
}
