% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calc_shr.R
\name{make_iter}
\alias{make_iter}
\title{Make Iterations Steps to Calculate Elasticity}
\usage{
make_iter(product_input, specs, n_points = 20)
}
\arguments{
\item{product_input}{Character. Product for which elasticity will be 
calculated.}

\item{specs}{Data frame. Table with products specification.}

\item{n_points}{Numeric. Number of points for which Max-Min range will be 
divided. Default `20`.}
}
\value{
Numeric Vector. A vector of prices to simulate the elasticity.
}
\description{
Function to calulates steps (price points) for elasticyty
  simulation.
}
