% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_elasticity_wrappers.R
\name{plot_e_share}
\alias{plot_e_share}
\title{Make Share Elasticity Plot}
\usage{
plot_e_share(iterated_df, product_input, product_output, ranges, e_share)
}
\arguments{
\item{iterated_df}{Data frame which is result of iter shares function.}

\item{product_input}{Product input}

\item{product_output}{Product output}

\item{ranges}{range of plot to put elasticity}

\item{e_share}{share elasticity value}
}
\value{
ggplot object
}
\description{
Elsticity Plot of shares.
}
