% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_elasticity_wrappers.R
\name{plot_e_sp}
\alias{plot_e_sp}
\title{Make Sales and Profit Plot}
\usage{
plot_e_sp(
  iterated_df,
  product_input,
  product_output,
  ranges,
  e_sales,
  e_ch_sales,
  e_profit
)
}
\arguments{
\item{iterated_df}{Data frame which is result of iter shares function.}

\item{product_input}{Product input.}

\item{product_output}{Product output.}

\item{ranges}{Range of plot to put elasticity.}

\item{e_sales}{Sales elasticity value.}

\item{e_ch_sales}{Channel sales elasticity value.}

\item{e_profit}{Profit elasticity value.}
}
\value{
`ggplot` object
}
\description{
Elasticity plot of sales, channel sales and profit.
}
