% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tbl_2_plot.R
\name{table_to_plot}
\alias{table_to_plot}
\title{Table for Plots}
\usage{
table_to_plot(
  data,
  specs,
  respid_key,
  product_names,
  current_prices,
  current_dm,
  current_cost,
  current_ms
)
}
\arguments{
\item{data}{Data frame with utilities.}

\item{specs}{Data frame with products specification.}

\item{respid_key}{Vector with Respondents ID to filter.}

\item{product_names}{Vector of product names.}

\item{current_prices}{Snapshot of prices from Assumptions tab.}

\item{current_dm}{Snapshot of distribution margin from Assumptions tab.}

\item{current_cost}{Snapshot of cost from Assumptions tab.}

\item{current_ms}{Snapshot of market size from Assumption tab.}
}
\value{
Data frame for plot functions.
}
\description{
Creates table to be used by plotting functions.
}
