% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_filter_df.R
\name{utils_filter_df}
\alias{utils_filter_df}
\title{Filter utils df to calculate shares based on choosen filters}
\usage{
utils_filter_df(data, respid_key)
}
\arguments{
\item{data}{Data frame with HB model utilities.}

\item{respid_key}{Vector of respondents IDs.}
}
\value{
Data frame with filtered records.
}
\description{
\code{utils_filter_df} filter utils df based on choosen filters.
}
