% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_optim_wrapper.R
\name{utils_optim_DT}
\alias{utils_optim_DT}
\title{DT Table of Optimized Shares, Sales and Profit}
\usage{
utils_optim_DT(
  product_input,
  product_output,
  data,
  specs,
  respid_key,
  product_names,
  current_prices,
  current_dm,
  current_cost,
  current_ms,
  n_points
)
}
\arguments{
\item{product_input}{Product for optimization.}

\item{data}{Data frame with hb utilities.}

\item{specs}{Data frame with specs.}

\item{respid_key}{Vector of respondents IDs.}

\item{product_names}{Vector of product names.}

\item{current_prices}{Vector of prices.}

\item{current_dm}{Vector of distribution margins.}

\item{current_cost}{Vector of costs.}

\item{current_ms}{Market size.}

\item{n_points}{Number of price points to optimize.}

\item{brand_output}{Product for optimization.}
}
\value{
Data frame with optimized shares, sales and profit.
}
\description{
Creates DT optimized table with shares, sales and profit.
}
