% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_calc_sample.R
\name{calc_sample}
\alias{calc_sample}
\title{Conjoint tools to calulate shares}
\usage{
calc_sample(respid_key)
}
\arguments{
\item{respid_key}{IDs from data frame with currently applied filters.}
}
\value{
Sample size.
}
\description{
\code{calc_shares} calculates shares, sales, channel sales and profit.
}
