% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calc_shr.R
\name{calc_shares}
\alias{calc_shares}
\title{Calculate Shares}
\usage{
calc_shares(
  data,
  specs,
  respid_key,
  product_names,
  current_prices,
  current_dm,
  current_cost,
  current_ms
)
}
\arguments{
\item{data}{Data frane with HB utilities.}

\item{specs}{Table with products specification.}

\item{respid_key}{Vector with Respondents ID to filter.}

\item{product_names}{Vector of product names.}

\item{current_prices}{Snapshot of prices from Assumptions tab.}

\item{current_dm}{Snapshot of distribution margin from Assumptions tab.}

\item{current_cost}{Snapshot of cost from Assumptions tab.}

\item{current_ms}{Snapshot of market size from Assumption tab.}
}
\value{
Data frame with calculated shares for a given price.
}
\description{
Calculates shares, sales, channel sales and profit of 
  all products in Specs.
}
