% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_main_html.R
\name{main_tab_html}
\alias{main_tab_html}
\title{Main Tab HTML}
\usage{
main_tab_html(simulator_name)
}
\arguments{
\item{simulator_name}{Name of the simulator to be displayed.}
}
\value{
Data frame with calculated shares by price step.
}
\description{
Create main tab html

\code{main_tab_html} generate_text.
}
