% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_null_to_zero.R
\name{null_to_zero}
\alias{null_to_zero}
\title{NULL to zero wrapper.}
\usage{
null_to_zero(data)
}
\arguments{
\item{data}{numeric or NULL value#'}
}
\value{
0 or \code{data} value.
}
\description{
Changes null values to 0.
}
