% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_shares_dt_wrapper.R
\name{utils_shares_DT}
\alias{utils_shares_DT}
\title{Shares and sales DT wrapper}
\usage{
utils_shares_DT(
  data,
  specs,
  respid_key,
  product_names,
  current_prices,
  current_dm,
  current_cost,
  current_ms
)
}
\arguments{
\item{data}{Data frame with hb utilities.}

\item{specs}{Data frame with specs.}

\item{respid_key}{Vector of respondents IDs.}

\item{product_names}{Vector of product names.}

\item{current_prices}{Vector of prices.}

\item{current_dm}{Vector of distribution margins.}

\item{current_cost}{Vector of costs.}

\item{current_ms}{Market size.}
}
\value{
Data frame with filtered records.
}
\description{
\code{utils_shares_DT} creates DT table with shares and sales.
}
